requirements.txt
fastapi
uvicorn
aiogram>=3.0.0b5
httpx
pydantic-settings
